+++ .storybook/main.js
@@ -2,6 +2,7 @@ const path = require('path');
 const less = require('less');
 const package = require('../package.json');
 const generate = require('generate-file-webpack-plugin');
+const vueConf = require('../vue.config');
 
 const cssLoaders = [
   'style-loader',
@@ -85,7 +86,16 @@ module.exports = {
     '../docs/**/*.stories.@(js|mdx)',
   ],
   addons: [
-    '@storybook/addon-docs',
+    {
+      name: '@storybook/addon-docs',
+      options: {
+        vueDocgenOptions: {
+          alias: {
+            '@': path.resolve(__dirname, '../')
+          },
+        },
+      }
+    },
     '@storybook/addon-controls',
     '@storybook/addon-actions',
     '@storybook/addon-links',