+++ components/keyboard_shortcut/keyboard_shortcut.stories.js
@@ -33,6 +33,9 @@ export default {
     docs: {
       page: DtKeyboardShortcutMdx,
     },
+    options: {
+      showPanel: true,
+    },
   },
 };
 
@@ -54,3 +57,4 @@ Default.args = {};
 
 export const Variants = VariantsTemplate.bind({});
 Variants.args = {};
+Variants.parameters = { controls: { disable: true }, actions: { disable: true }, options: { showPanel: false } };