+++ components/link/link.stories.js
@@ -112,6 +112,9 @@ export default {
     docs: {
       page: DtLinkMdx,
     },
+    options: {
+      showPanel: true,
+    },
   },
 };
 
@@ -125,3 +128,4 @@ Default.args = {};
 
 export const Variants = VariantsTemplate.bind({});
 Variants.args = {};
+Variants.parameters = { controls: { disable: true }, actions: { disable: true }, options: { showPanel: false } };