+++ components/notice/notice.stories.js
@@ -95,6 +95,9 @@ export default {
     docs: {
       page: DtNoticeMdx,
     },
+    options: {
+      showPanel: true,
+    },
   },
 };