+++ recipes/notices/top_banner_info/top_banner_info_constants.js
@@ -1,0 +1,6 @@
+export const COLOR_CODES = ['green400', 'green200', 'red200', 'red100', 'yellow200',
+  'yellow400', 'yellow050', 'black100', 'white'];
+
+export default {
+  COLOR_CODES,
+};